Crate mysqlclient_sys

Reexports

pub use self::_bindgen_ty_2 as idtype_t;

Structs

PSI_thread

__BindgenUnionField

__darwin_fp_control

__darwin_fp_status

__darwin_i386_avx_state

__darwin_i386_exception_state

__darwin_i386_float_state

__darwin_i386_thread_state

__darwin_mcontext32

__darwin_mcontext64

__darwin_mcontext_avx32

__darwin_mcontext_avx64

__darwin_mmst_reg

__darwin_pthread_handler_rec

__darwin_sigaltstack

__darwin_ucontext

__darwin_x86_avx_state64

__darwin_x86_debug_state32

__darwin_x86_debug_state64

__darwin_x86_exception_state64

__darwin_x86_float_state64

__darwin_x86_thread_state64

__darwin_xmm_reg

__sigaction

__sigaction_u

__siginfo

__va_list_tag

_bindgen_ty_1

_bindgen_ty_3

_bindgen_ty_4

_bindgen_ty_5

_opaque_pthread_attr_t

_opaque_pthread_cond_t

_opaque_pthread_condattr_t

_opaque_pthread_mutex_t

_opaque_pthread_mutexattr_t

_opaque_pthread_once_t

_opaque_pthread_rwlock_t

_opaque_pthread_rwlockattr_t

_opaque_pthread_t

character_set

embedded_query_result

fd_set

proc_rlimit_control_wakeupmon

rand_struct

rlimit

rusage

rusage_info_v0

rusage_info_v1

rusage_info_v2

rusage_info_v3

sigaction

sigevent

sigstack

sigval

sigvec

st_list

st_mem_root

st_mysql

st_mysql_bind

st_mysql_charset_info_st

st_mysql_client_plugin

st_mysql_client_plugin_AUTHENTICATION

st_mysql_data

st_mysql_field

st_mysql_methods

st_mysql_options

st_mysql_options__bindgen_ty_1

st_mysql_options_extention

st_mysql_options_st_dynamic_array

st_mysql_res

st_mysql_rows

st_mysql_stmt

st_mysql_stmt_extension

st_mysql_time

st_net

st_plugin_vio
Provides plugin access to communication channel

st_plugin_vio_info

st_typelib

st_udf_args

st_udf_init

st_used_mem

st_vio

timeval

wait

wait__bindgen_ty_1

wait__bindgen_ty_2

Enums

Item_result

_bindgen_ty_2

enum_cursor_type

enum_field_types

enum_mysql_set_option

enum_mysql_stmt_state

enum_mysql_timestamp_type

enum_server_command
@enum enum_server_command @brief You should add new commands to the end of this list, otherwise old servers won't be able to handle them as 'unsupported'.

enum_session_state_type

enum_stmt_attr_type

mysql_enum_shutdown_level

mysql_option

mysql_protocol_type

mysql_ssl_mode

mysql_status

st_plugin_vio_info__bindgen_ty_1

Constants

ALLOC_MAX_BLOCK_TO_DROP

ALLOC_MAX_BLOCK_USAGE_BEFORE_DROP

AUTH_PLUGIN_DATA_PART_1_LENGTH

AUTO_INCREMENT_FLAG

BIG_ENDIAN

BINARY_FLAG

BINCMP_FLAG

BLOB_FLAG

BUS_ADRALN

BUS_ADRERR

BUS_NOOP

BUS_OBJERR

BYTE_ORDER

CAN_CLIENT_COMPRESS

CLD_CONTINUED

CLD_DUMPED

CLD_EXITED

CLD_KILLED

CLD_NOOP

CLD_STOPPED

CLD_TRAPPED

CLIENT_CAN_HANDLE_EXPIRED_PASSWORDS

CLIENT_COMPRESS

CLIENT_CONNECT_ATTRS

CLIENT_CONNECT_WITH_DB

CLIENT_DEPRECATE_EOF

CLIENT_FOUND_ROWS

CLIENT_IGNORE_SIGPIPE

CLIENT_IGNORE_SPACE

CLIENT_INTERACTIVE

CLIENT_LOCAL_FILES

CLIENT_LONG_FLAG

CLIENT_LONG_PASSWORD

CLIENT_MULTI_QUERIES

CLIENT_MULTI_RESULTS

CLIENT_MULTI_STATEMENTS

CLIENT_NET_READ_TIMEOUT

CLIENT_NET_WRITE_TIMEOUT

CLIENT_NO_SCHEMA

CLIENT_ODBC

CLIENT_PLUGIN_AUTH

CLIENT_PLUGIN_AUTH_LENENC_CLIENT_DATA

CLIENT_PROTOCOL_41

CLIENT_PS_MULTI_RESULTS

CLIENT_REMEMBER_OPTIONS

CLIENT_RESERVED

CLIENT_RESERVED2

CLIENT_SESSION_TRACK

CLIENT_SSL

CLIENT_SSL_VERIFY_SERVER_CERT

CLIENT_TRANSACTIONS

COLUMN_COMMENT_MAXLEN

COMP_HEADER_SIZE

CPUMON_MAKE_FATAL

CR_AUTH_HANDSHAKE

CR_AUTH_PLUGIN_ERROR

CR_AUTH_USER_CREDENTIALS

CR_ERROR

CR_OK

CR_OK_HANDSHAKE_COMPLETE

DATETIME_MAX_DECIMALS

ENUM_FLAG

EXIT_FAILURE

EXIT_SUCCESS

EXPLICIT_NULL_FLAG

FD_SETSIZE

FIELD_FLAGS_COLUMN_FORMAT

FIELD_FLAGS_COLUMN_FORMAT_MASK

FIELD_FLAGS_STORAGE_MEDIA

FIELD_FLAGS_STORAGE_MEDIA_MASK

FIELD_IN_ADD_INDEX

FIELD_IN_PART_FUNC_FLAG

FIELD_IS_DROPPED

FIELD_IS_RENAMED

FILENAME_CHARSET_MBMAXLEN

FIND_TYPE_ALLOW_NUMBER

FIND_TYPE_BASIC

FIND_TYPE_COMMA_TERM

FIND_TYPE_NO_OVERWRITE

FIND_TYPE_NO_PREFIX

FPE_FLTDIV

FPE_FLTINV

FPE_FLTOVF

FPE_FLTRES

FPE_FLTSUB

FPE_FLTUND

FPE_INTDIV

FPE_INTOVF

FPE_NOOP

FP_CHOP

FP_PREC_24B

FP_PREC_53B

FP_PREC_64B

FP_RND_DOWN

FP_RND_NEAR

FP_RND_UP

FP_STATE_BYTES

FRM_GCOL_HEADER_SIZE

FRM_VER

GENERATED_COLUMN_EXPRESSION_MAXLEN

GET_FIXED_FIELDS_FLAG

GROUP_FLAG

HOSTNAME_LENGTH

ILL_BADSTK

ILL_COPROC

ILL_ILLADR

ILL_ILLOPC

ILL_ILLOPN

ILL_ILLTRP

ILL_NOOP

ILL_PRVOPC

ILL_PRVREG

INDEX_COMMENT_MAXLEN

INT16_MAX

INT16_MIN

INT32_MAX

INT32_MIN

INT64_MAX

INT64_MIN

INT8_MAX

INT8_MIN

INTMAX_MAX

INTMAX_MIN

INTPTR_MAX

INTPTR_MIN

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST32_MAX

INT_FAST32_MIN

INT_FAST64_MAX

INT_FAST64_MIN

INT_FAST8_MAX

INT_FAST8_MIN

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST32_MAX

INT_LEAST32_MIN

INT_LEAST64_MAX

INT_LEAST64_MIN

INT_LEAST8_MAX

INT_LEAST8_MIN

IOPOL_APPLICATION

IOPOL_DEFAULT

IOPOL_IMPORTANT

IOPOL_NORMAL

IOPOL_PASSIVE

IOPOL_SCOPE_DARWIN_BG

IOPOL_SCOPE_PROCESS

IOPOL_SCOPE_THREAD

IOPOL_STANDARD

IOPOL_THROTTLE

IOPOL_TYPE_DISK

IOPOL_UTILITY

LIBMYSQL_VERSION

LIBMYSQL_VERSION_ID

LITTLE_ENDIAN

LOCAL_HOST

LOCAL_HOST_NAMEDPIPE

LOCAL_INFILE_ERROR_LEN

MAX_BIGINT_WIDTH

MAX_BLOB_WIDTH

MAX_CHAR_WIDTH

MAX_INT_WIDTH

MAX_MEDIUMINT_WIDTH

MAX_PACKET_LENGTH

MAX_SMALLINT_WIDTH

MAX_TINYINT_WIDTH

MINSIGSTKSZ

MULTIPLE_KEY_FLAG

MYSQL_AUTODETECT_CHARSET_NAME

MYSQL_BASE_VERSION

MYSQL_CLIENT_AUTHENTICATION_PLUGIN

MYSQL_CLIENT_AUTHENTICATION_PLUGIN_INTERFACE_VERSION

MYSQL_CLIENT_MAX_PLUGINS

MYSQL_CLIENT_TRACE_PLUGIN

MYSQL_CLIENT_TRACE_PLUGIN_INTERFACE_VERSION

MYSQL_CLIENT_reserved1

MYSQL_CLIENT_reserved2

MYSQL_COMPILATION_COMMENT

MYSQL_CONFIG_NAME

MYSQL_DATA_TRUNCATED

MYSQL_ERRMSG_SIZE

MYSQL_LONG_DATA_HEADER

MYSQL_NO_DATA

MYSQL_PORT

MYSQL_PORT_DEFAULT

MYSQL_SERVER_SUFFIX_DEF

MYSQL_SERVER_VERSION

MYSQL_STMT_HEADER

MYSQL_UNIX_ADDR

MYSQL_USERNAME_LENGTH

MYSQL_VERSION_ID

NAME_CHAR_LEN

NAME_LEN

NBBY

NET_HEADER_SIZE

NET_READ_TIMEOUT

NET_WAIT_TIMEOUT

NET_WRITE_TIMEOUT

NOT_FIXED_DEC

NOT_NULL_FLAG

NO_DEFAULT_VALUE_FLAG

NSIG

NUM_FLAG

ONLY_KILL_QUERY

ON_UPDATE_NOW_FLAG

PART_KEY_FLAG

PDP_ENDIAN

POLL_ERR

POLL_HUP

POLL_IN

POLL_MSG

POLL_OUT

POLL_PRI

PRIO_DARWIN_BG

PRIO_DARWIN_NONUI

PRIO_DARWIN_PROCESS

PRIO_DARWIN_THREAD

PRIO_MAX

PRIO_MIN

PRIO_PGRP

PRIO_PROCESS

PRIO_USER

PRI_KEY_FLAG

PROTOCOL_VERSION

PROXY_FLAG

PSI_FLAG_GLOBAL

PSI_FLAG_MUTABLE

PSI_FLAG_STAGE_PROGRESS

PSI_FLAG_THREAD

PSI_FLAG_TRANSFER

PSI_INSTRUMENT_ME

PSI_NOT_INSTRUMENTED

PSI_RWLOCK_FLAG_SX

PTRDIFF_MAX

PTRDIFF_MIN

P_ALL

P_PGID

P_PID

RAND_MAX

REFRESH_BINARY_LOG

REFRESH_DES_KEY_FILE

REFRESH_ENGINE_LOG

REFRESH_ERROR_LOG

REFRESH_FAST

REFRESH_FOR_EXPORT

REFRESH_GENERAL_LOG

REFRESH_GRANT

REFRESH_HOSTS

REFRESH_LOG

REFRESH_MASTER

REFRESH_OPTIMIZER_COSTS

REFRESH_QUERY_CACHE

REFRESH_QUERY_CACHE_FREE

REFRESH_READ_LOCK

REFRESH_RELAY_LOG

REFRESH_SLAVE

REFRESH_SLOW_LOG

REFRESH_STATUS

REFRESH_TABLES

REFRESH_THREADS

REFRESH_USER_RESOURCES

RLIMIT_AS

RLIMIT_CORE

RLIMIT_CPU

RLIMIT_CPU_USAGE_MONITOR

RLIMIT_DATA

RLIMIT_FSIZE

RLIMIT_MEMLOCK

RLIMIT_NOFILE

RLIMIT_NPROC

RLIMIT_RSS

RLIMIT_STACK

RLIMIT_THREAD_CPULIMITS

RLIMIT_WAKEUPS_MONITOR

RLIM_NLIMITS

RSIZE_MAX

RUSAGE_CHILDREN

RUSAGE_INFO_CURRENT

RUSAGE_INFO_V0

RUSAGE_INFO_V1

RUSAGE_INFO_V2

RUSAGE_INFO_V3

RUSAGE_SELF

SA_64REGSET

SA_NOCLDSTOP

SA_NOCLDWAIT

SA_NODEFER

SA_ONSTACK

SA_RESETHAND

SA_RESTART

SA_SIGINFO

SA_USERSPACE_MASK

SA_USERTRAMP

SCRAMBLED_PASSWORD_CHAR_LENGTH

SCRAMBLE_LENGTH

SEGV_ACCERR

SEGV_MAPERR

SEGV_NOOP

SERVER_MORE_RESULTS_EXISTS

SERVER_PS_OUT_PARAMS

SERVER_QUERY_NO_GOOD_INDEX_USED

SERVER_QUERY_NO_INDEX_USED

SERVER_QUERY_WAS_SLOW

SERVER_SESSION_STATE_CHANGED

SERVER_STATUS_AUTOCOMMIT

SERVER_STATUS_CLEAR_SET

SERVER_STATUS_CURSOR_EXISTS

SERVER_STATUS_DB_DROPPED

SERVER_STATUS_IN_TRANS

SERVER_STATUS_IN_TRANS_READONLY

SERVER_STATUS_LAST_ROW_SENT

SERVER_STATUS_METADATA_CHANGED

SERVER_STATUS_NO_BACKSLASH_ESCAPES

SERVER_VERSION_LENGTH

SET_FLAG

SIGABRT

SIGALRM

SIGBUS

SIGCHLD

SIGCONT

SIGEMT

SIGEV_NONE

SIGEV_SIGNAL

SIGEV_THREAD

SIGFPE

SIGHUP

SIGILL

SIGINFO

SIGINT

SIGIO

SIGIOT

SIGKILL

SIGPIPE

SIGPROF

SIGQUIT

SIGSEGV

SIGSTKSZ

SIGSTOP

SIGSYS

SIGTERM

SIGTRAP

SIGTSTP

SIGTTIN

SIGTTOU

SIGURG

SIGUSR1

SIGUSR2

SIGVTALRM

SIGWINCH

SIGXCPU

SIGXFSZ

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIG_BLOCK

SIG_SETMASK

SIG_UNBLOCK

SIZE_MAX

SI_ASYNCIO

SI_MESGQ

SI_QUEUE

SI_TIMER

SI_USER

SQLSTATE_LENGTH

SS_DISABLE

SS_ONSTACK

SV_INTERRUPT

SV_NOCLDSTOP

SV_NODEFER

SV_ONSTACK

SV_RESETHAND

SV_SIGINFO

SYSTEM_CHARSET_MBMAXLEN

SYS_SCHEMA_VERSION

TABLE_COMMENT_INLINE_MAXLEN

TABLE_COMMENT_MAXLEN

TABLE_PARTITION_COMMENT_MAXLEN

TIMESTAMP_FLAG

TRAP_BRKPT

TRAP_TRACE

UINT16_MAX

UINT32_MAX

UINT64_MAX

UINT8_MAX

UINTMAX_MAX

UINTPTR_MAX

UINT_FAST16_MAX

UINT_FAST32_MAX

UINT_FAST64_MAX

UINT_FAST8_MAX

UINT_LEAST16_MAX

UINT_LEAST32_MAX

UINT_LEAST64_MAX

UINT_LEAST8_MAX

UNIQUE_FLAG

UNIQUE_KEY_FLAG

UNSIGNED_FLAG

USERNAME_CHAR_LENGTH

USERNAME_CHAR_LENGTH_STR

USERNAME_LENGTH

USER_HOST_BUFF_SIZE

WAIT_ANY

WAIT_MYPGRP

WAKEMON_DISABLE

WAKEMON_ENABLE

WAKEMON_GET_PARAMS

WAKEMON_MAKE_FATAL

WAKEMON_SET_DEFAULTS

WCONTINUED

WCOREFLAG

WEXITED

WINT_MAX

WINT_MIN

WNOHANG

WNOWAIT

WSTOPPED

WUNTRACED

ZEROFILL_FLAG

_DARWIN_FEATURE_64_BIT_INODE

_DARWIN_FEATURE_ONLY_UNIX_CONFORMANCE

_DARWIN_FEATURE_UNIX_CONFORMANCE

_FORTIFY_SOURCE

_I386_SIGNAL_H_

_QUAD_HIGHWORD

_QUAD_LOWWORD

_RLIMIT_POSIX_FLAG

_WSTOPPED

__DARWIN_64_BIT_INO_T

__DARWIN_BIG_ENDIAN

__DARWIN_BYTE_ORDER

__DARWIN_C_ANSI

__DARWIN_C_FULL

__DARWIN_C_LEVEL

__DARWIN_FD_SETSIZE

__DARWIN_LITTLE_ENDIAN

__DARWIN_NBBY

__DARWIN_NON_CANCELABLE

__DARWIN_NSIG

__DARWIN_ONLY_64_BIT_INO_T

__DARWIN_ONLY_UNIX_CONFORMANCE

__DARWIN_ONLY_VERS_1050

__DARWIN_PDP_ENDIAN

__DARWIN_SUF_1050

__DARWIN_SUF_64_BIT_INO_T

__DARWIN_SUF_EXTSN

__DARWIN_UNIX03

__DARWIN_VERS_1050

__DARWIN_WCHAR_MIN

__GNUC_VA_LIST

__IPHONE_2_0

__IPHONE_2_1

__IPHONE_2_2

__IPHONE_3_0

__IPHONE_3_1

__IPHONE_3_2

__IPHONE_4_0

__IPHONE_4_1

__IPHONE_4_2

__IPHONE_4_3

__IPHONE_5_0

__IPHONE_5_1

__IPHONE_6_0

__IPHONE_6_1

__IPHONE_7_0

__IPHONE_7_1

__IPHONE_8_0

__MAC_10_0

__MAC_10_1

__MAC_10_10

__MAC_10_10_2

__MAC_10_10_3

__MAC_10_2

__MAC_10_3

__MAC_10_4

__MAC_10_5

__MAC_10_6

__MAC_10_7

__MAC_10_8

__MAC_10_9

__MAC_OS_X_VERSION_MAX_ALLOWED

__PTHREAD_ATTR_SIZE__

__PTHREAD_CONDATTR_SIZE__

__PTHREAD_COND_SIZE__

__PTHREAD_MUTEXATTR_SIZE__

__PTHREAD_MUTEX_SIZE__

__PTHREAD_ONCE_SIZE__

__PTHREAD_RWLOCKATTR_SIZE__

__PTHREAD_RWLOCK_SIZE__

__PTHREAD_SIZE__

__STDC_WANT_LIB_EXT1__

__WORDSIZE

st_plugin_vio_info_MYSQL_VIO_INVALID

st_plugin_vio_info_MYSQL_VIO_MEMORY

st_plugin_vio_info_MYSQL_VIO_PIPE

st_plugin_vio_info_MYSQL_VIO_SOCKET

st_plugin_vio_info_MYSQL_VIO_TCP

Statics

__mb_cur_max

mysql_port

mysql_unix_port

sql_protocol_typelib

suboptarg

Functions

_Exit

a64l

abort

abs

alloca

arc4random

arc4random_addrandom

arc4random_buf

arc4random_stir

arc4random_uniform

atexit

atexit_b

atof

atoi

atol

atoll

bsearch

bsearch_b

calloc

cgetcap

cgetclose

cgetent

cgetfirst

cgetmatch

cgetnext

cgetnum

cgetset

cgetstr

cgetustr

check_scramble

check_scramble_323

copy_typelib

create_random_string

daemon

devname

devname_r

div

drand48

ecvt

erand48

exit

fcvt

find_set_from_flags

find_type

find_type_or_exit

find_typeset

free

gcvt

get_salt_from_password

get_salt_from_password_323

get_tty_password

get_type

getbsize

getenv

getiopolicy_np

getloadavg

getpriority

getprogname

getrlimit

getrusage

getsubopt

grantpt

hash_password

heapsort

heapsort_b

initstate

jrand48

l64a

labs

lcong48

ldiv

list_add

list_cons

list_delete

list_free

list_length

list_reverse

list_walk

llabs

lldiv

lrand48

make_password_from_salt

make_password_from_salt_323

make_scrambled_password

make_scrambled_password_323

make_type

malloc

mblen

mbstowcs

mbtowc

mergesort

mergesort_b

mkstemp

mktemp

mrand48

my_net_init

my_net_local_init

my_net_read

my_net_write

my_rnd

my_thread_end

my_thread_init

myodbc_remove_escape

mysql_affected_rows

mysql_autocommit

mysql_change_user

mysql_character_set_name

mysql_client_find_plugin
finds an already loaded plugin by name, or loads it, if necessary

mysql_client_register_plugin
adds a plugin structure to the list of loaded plugins

mysql_close

mysql_commit

mysql_data_seek

mysql_debug

mysql_dump_debug_info

mysql_embedded

mysql_eof

mysql_errno

mysql_errno_to_sqlstate

mysql_error

mysql_escape_string

mysql_fetch_field

mysql_fetch_field_direct

mysql_fetch_fields

mysql_fetch_lengths

mysql_fetch_row

mysql_field_count

mysql_field_seek

mysql_field_tell

mysql_free_result

mysql_get_character_set_info

mysql_get_client_info

mysql_get_client_version

mysql_get_host_info

mysql_get_option

mysql_get_proto_info

mysql_get_server_info

mysql_get_server_version

mysql_get_ssl_cipher

mysql_hex_string

mysql_info

mysql_init

mysql_insert_id

mysql_kill

mysql_list_dbs

mysql_list_fields

mysql_list_processes

mysql_list_tables

mysql_load_plugin
loads a plugin and initializes it

mysql_load_plugin_v
loads a plugin and initializes it, taking va_list as an argument

mysql_more_results

mysql_next_result

mysql_num_fields

mysql_num_rows

mysql_options

mysql_options4

mysql_ping

mysql_plugin_options
set plugin options

mysql_query

mysql_read_query_result

mysql_real_connect

mysql_real_escape_string

mysql_real_escape_string_quote

mysql_real_query

mysql_refresh

mysql_reset_connection

mysql_rollback

mysql_row_seek

mysql_row_tell

mysql_select_db

mysql_send_query

mysql_server_end

mysql_server_init

mysql_session_track_get_first

mysql_session_track_get_next

mysql_set_character_set

mysql_set_local_infile_default

mysql_set_local_infile_handler

mysql_set_server_option

mysql_shutdown

mysql_sqlstate

mysql_ssl_set

mysql_stat

mysql_stmt_affected_rows

mysql_stmt_attr_get

mysql_stmt_attr_set

mysql_stmt_bind_param

mysql_stmt_bind_result

mysql_stmt_close

mysql_stmt_data_seek

mysql_stmt_errno

mysql_stmt_error

mysql_stmt_execute

mysql_stmt_fetch

mysql_stmt_fetch_column

mysql_stmt_field_count

mysql_stmt_free_result

mysql_stmt_init

mysql_stmt_insert_id

mysql_stmt_next_result

mysql_stmt_num_rows

mysql_stmt_param_count

mysql_stmt_param_metadata

mysql_stmt_prepare

mysql_stmt_reset

mysql_stmt_result_metadata

mysql_stmt_row_seek

mysql_stmt_row_tell

mysql_stmt_send_long_data

mysql_stmt_sqlstate

mysql_stmt_store_result

mysql_store_result

mysql_thread_end

mysql_thread_id

mysql_thread_init

mysql_thread_safe

mysql_use_result

mysql_warning_count

net_claim_memory_ownership

net_clear

net_end

net_flush

net_realloc

net_write_command

net_write_packet

nrand48

octet2hex

posix_memalign

posix_openpt

psort

psort_b

psort_r

ptsname

putenv

qsort

qsort_b

qsort_r

radixsort

rand

rand_r

random

randominit

realloc

reallocf

realpath

scramble

scramble_323

seed48

setenv

setiopolicy_np

setkey

setpriority

setprogname

setrlimit

setstate

signal

sradixsort

srand

srand48

sranddev

srandom

srandomdev

strtod

strtof

strtol

strtold

strtoll

strtoq

strtoul

strtoull

strtouq

system

unlockpt

unsetenv

valloc

wait

wait3

wait4

waitid

waitpid

wcstombs

wctomb

Type Definitions

EMBEDDED_QUERY_RESULT

LIST

MEM_ROOT

MYSQL

MYSQL_BIND

MYSQL_DATA

MYSQL_FIELD

MYSQL_FIELD_OFFSET

MYSQL_PLUGIN_VIO

MYSQL_PLUGIN_VIO_INFO

MYSQL_RES

MYSQL_ROW

MYSQL_ROWS

MYSQL_ROW_OFFSET

MYSQL_STMT

MYSQL_TIME

MY_CHARSET_INFO

NET

PSI_memory_key
Instrumented memory key. To instrument memory, a memory key must be obtained using @c register_memory. Using a zero key always disable the instrumentation.

TYPELIB

UDF_ARGS

UDF_INIT

USED_MEM

Vio

__builtin_va_list

__darwin_blkcnt_t

__darwin_blksize_t

__darwin_clock_t

__darwin_ct_rune_t

__darwin_dev_t

__darwin_fp_control_t

__darwin_fp_status_t

__darwin_fsblkcnt_t

__darwin_fsfilcnt_t

__darwin_gid_t

__darwin_id_t

__darwin_ino64_t

__darwin_ino_t

__darwin_intptr_t

__darwin_mach_port_name_t

__darwin_mach_port_t

__darwin_mbstate_t

__darwin_mode_t

__darwin_natural_t

__darwin_nl_item

__darwin_off_t

__darwin_pid_t

__darwin_pthread_attr_t

__darwin_pthread_cond_t

__darwin_pthread_condattr_t

__darwin_pthread_key_t

__darwin_pthread_mutex_t

__darwin_pthread_mutexattr_t

__darwin_pthread_once_t

__darwin_pthread_rwlock_t

__darwin_pthread_rwlockattr_t

__darwin_pthread_t

__darwin_ptrdiff_t

__darwin_rune_t

__darwin_sigset_t

__darwin_size_t

__darwin_socklen_t

__darwin_ssize_t

__darwin_suseconds_t

__darwin_time_t

__darwin_uid_t

__darwin_useconds_t

__darwin_uuid_string_t

__darwin_uuid_t

__darwin_va_list

__darwin_wchar_t

__darwin_wctrans_t

__darwin_wctype_t

__darwin_wint_t

__gnuc_va_list

__int16_t

__int32_t

__int64_t

__int8_t

__mbstate_t

__uint16_t

__uint32_t

__uint64_t

__uint8_t

blkcnt_t

blksize_t

caddr_t

clock_t

ct_rune_t

daddr_t

dev_t

div_t

errno_t

fd_mask

fixpt_t

fsblkcnt_t

fsfilcnt_t

gid_t

id_t

in_addr_t

in_port_t

ino64_t

ino_t

int_fast16_t

int_fast32_t

int_fast64_t

int_fast8_t

int_least16_t

int_least32_t

int_least64_t

int_least8_t

intmax_t

key_t

ldiv_t

list_walk_action

lldiv_t

mcontext_t

mode_t

my_bool

my_socket

my_ulonglong

nlink_t

off_t

pid_t

pthread_attr_t

pthread_cond_t

pthread_condattr_t

pthread_key_t

pthread_mutex_t

pthread_mutexattr_t

pthread_once_t

pthread_rwlock_t

pthread_rwlockattr_t

pthread_t

qaddr_t

quad_t

register_t

rlim_t

rsize_t

rune_t

rusage_info_current

rusage_info_t

segsz_t

sig_atomic_t

sig_t

siginfo_t

sigset_t

stack_t

suseconds_t

swblk_t

syscall_arg_t

time_t

u_char

u_int

u_int16_t

u_int32_t

u_int64_t

u_int8_t

u_long

u_quad_t

u_short

ucontext_t

uid_t

uint

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_fast8_t

uint_least16_t

uint_least32_t

uint_least64_t

uint_least8_t

uintmax_t

useconds_t

user_addr_t

user_long_t

user_off_t

user_size_t

user_ssize_t

user_time_t

user_ulong_t

ushort

va_list

wchar_t