mysqlclient_sys

Function mysql_default_compression_level

pub unsafe extern "C" fn mysql_default_compression_level(
    algorithm: enum_compression_algorithm,
) -> c_uint

Get default compression level corresponding to a given compression method.

@param algorithm Compression Method. Possible values are zlib or zstd.

@return an unsigned int representing default compression level. 6 is the default compression level for zlib and 3 is the default compression level for zstd.