Crate mysqlclient_sys

Structs

__BindgenUnionField

__va_list_tag

character_set

embedded_query_result

st_list

st_mem_root

st_mysql

st_mysql_bind

st_mysql_charset_info_st

st_mysql_client_plugin

st_mysql_data

st_mysql_field

st_mysql_methods

st_mysql_options

st_mysql_options__bindgen_ty_1

st_mysql_options_extention

st_mysql_options_st_dynamic_array

st_mysql_res

st_mysql_rows

st_mysql_stmt

st_mysql_stmt_extension

st_mysql_time

st_net

st_plugin_vio
Provides plugin access to communication channel

st_plugin_vio_info

st_used_mem

st_vio

Enums

enum_field_types

enum_mysql_set_option

enum_mysql_stmt_state

enum_mysql_timestamp_type

enum_session_state_type

enum_stmt_attr_type

mysql_enum_shutdown_level

mysql_option

mysql_protocol_type

mysql_ssl_mode

mysql_status

st_plugin_vio_info__bindgen_ty_1

Constants

MYSQL_AUTODETECT_CHARSET_NAME

MYSQL_BASE_VERSION

MYSQL_CLIENT_AUTHENTICATION_PLUGIN

MYSQL_CLIENT_AUTHENTICATION_PLUGIN_INTERFACE_VERSION

MYSQL_CLIENT_MAX_PLUGINS

MYSQL_CLIENT_TRACE_PLUGIN

MYSQL_CLIENT_TRACE_PLUGIN_INTERFACE_VERSION

MYSQL_CLIENT_reserved1

MYSQL_CLIENT_reserved2

MYSQL_COMPILATION_COMMENT

MYSQL_CONFIG_NAME

MYSQL_DATA_TRUNCATED

MYSQL_ERRMSG_SIZE

MYSQL_LONG_DATA_HEADER

MYSQL_NO_DATA

MYSQL_PORT

MYSQL_PORT_DEFAULT

MYSQL_SERVER_SUFFIX_DEF

MYSQL_SERVER_VERSION

MYSQL_STMT_HEADER

MYSQL_UNIX_ADDR

MYSQL_USERNAME_LENGTH

MYSQL_VERSION_ID

st_plugin_vio_info_MYSQL_VIO_INVALID

st_plugin_vio_info_MYSQL_VIO_MEMORY

st_plugin_vio_info_MYSQL_VIO_PIPE

st_plugin_vio_info_MYSQL_VIO_SOCKET

st_plugin_vio_info_MYSQL_VIO_TCP

Functions

mysql_affected_rows⚠

mysql_autocommit⚠

mysql_change_user⚠

mysql_character_set_name⚠

mysql_client_find_plugin⚠
finds an already loaded plugin by name, or loads it, if necessary

mysql_client_register_plugin⚠
adds a plugin structure to the list of loaded plugins

mysql_close⚠

mysql_commit⚠

mysql_data_seek⚠

mysql_debug⚠

mysql_dump_debug_info⚠

mysql_embedded⚠

mysql_eof⚠

mysql_errno⚠

mysql_errno_to_sqlstate⚠

mysql_error⚠

mysql_escape_string⚠

mysql_fetch_field⚠

mysql_fetch_field_direct⚠

mysql_fetch_fields⚠

mysql_fetch_lengths⚠

mysql_fetch_row⚠

mysql_field_count⚠

mysql_field_seek⚠

mysql_field_tell⚠

mysql_free_result⚠

mysql_get_character_set_info⚠

mysql_get_client_info⚠

mysql_get_client_version⚠

mysql_get_host_info⚠

mysql_get_option⚠

mysql_get_proto_info⚠

mysql_get_server_info⚠

mysql_get_server_version⚠

mysql_get_ssl_cipher⚠

mysql_hex_string⚠

mysql_info⚠

mysql_init⚠

mysql_insert_id⚠

mysql_kill⚠

mysql_list_dbs⚠

mysql_list_fields⚠

mysql_list_processes⚠

mysql_list_tables⚠

mysql_load_plugin⚠
loads a plugin and initializes it

mysql_load_plugin_v⚠
loads a plugin and initializes it, taking va_list as an argument

mysql_more_results⚠

mysql_next_result⚠

mysql_num_fields⚠

mysql_num_rows⚠

mysql_options⚠

mysql_options4⚠

mysql_ping⚠

mysql_plugin_options⚠
set plugin options

mysql_query⚠

mysql_read_query_result⚠

mysql_real_connect⚠

mysql_real_escape_string⚠

mysql_real_escape_string_quote⚠

mysql_real_query⚠

mysql_refresh⚠

mysql_reset_connection⚠

mysql_rollback⚠

mysql_row_seek⚠

mysql_row_tell⚠

mysql_select_db⚠

mysql_send_query⚠

mysql_server_end⚠

mysql_server_init⚠

mysql_session_track_get_first⚠

mysql_session_track_get_next⚠

mysql_set_character_set⚠

mysql_set_local_infile_default⚠

mysql_set_local_infile_handler⚠

mysql_set_server_option⚠

mysql_shutdown⚠

mysql_sqlstate⚠

mysql_ssl_set⚠

mysql_stat⚠

mysql_stmt_affected_rows⚠

mysql_stmt_attr_get⚠

mysql_stmt_attr_set⚠

mysql_stmt_bind_param⚠

mysql_stmt_bind_result⚠

mysql_stmt_close⚠

mysql_stmt_data_seek⚠

mysql_stmt_errno⚠

mysql_stmt_error⚠

mysql_stmt_execute⚠

mysql_stmt_fetch⚠

mysql_stmt_fetch_column⚠

mysql_stmt_field_count⚠

mysql_stmt_free_result⚠

mysql_stmt_init⚠

mysql_stmt_insert_id⚠

mysql_stmt_next_result⚠

mysql_stmt_num_rows⚠

mysql_stmt_param_count⚠

mysql_stmt_param_metadata⚠

mysql_stmt_prepare⚠

mysql_stmt_reset⚠

mysql_stmt_result_metadata⚠

mysql_stmt_row_seek⚠

mysql_stmt_row_tell⚠

mysql_stmt_send_long_data⚠

mysql_stmt_sqlstate⚠

mysql_stmt_store_result⚠

mysql_store_result⚠

mysql_thread_end⚠

mysql_thread_id⚠

mysql_thread_init⚠

mysql_thread_safe⚠

mysql_use_result⚠

mysql_warning_count⚠

Type Definitions

LIST

MEM_ROOT

MYSQL

MYSQL_BIND

MYSQL_DATA

MYSQL_FIELD

MYSQL_FIELD_OFFSET

MYSQL_PLUGIN_VIO

MYSQL_PLUGIN_VIO_INFO

MYSQL_RES

MYSQL_ROW

MYSQL_ROWS

MYSQL_ROW_OFFSET

MYSQL_STMT

MYSQL_TIME

MY_CHARSET_INFO

NET

PSI_memory_key
Instrumented memory key. To instrument memory, a memory key must be obtained using @c register_memory. Using a zero key always disable the instrumentation.

USED_MEM

Vio

__builtin_va_list

__darwin_size_t

my_bool

my_socket

my_ulonglong

va_list