Struct mysqlclient_sys::st_net

#[repr(C)]
pub struct st_net {

    pub vio: *mut Vio,
    pub buff: *mut c_uchar,
    pub buff_end: *mut c_uchar,
    pub write_pos: *mut c_uchar,
    pub read_pos: *mut c_uchar,
    pub fd: my_socket,
    pub remain_in_buf: c_ulong,
    pub length: c_ulong,
    pub buf_length: c_ulong,
    pub where_b: c_ulong,
    pub max_packet: c_ulong,
    pub max_packet_size: c_ulong,
    pub pkt_nr: c_uint,
    pub compress_pkt_nr: c_uint,
    pub write_timeout: c_uint,
    pub read_timeout: c_uint,
    pub retry_count: c_uint,
    pub fcntl: c_int,
    pub return_status: *mut c_uint,
    pub reading_or_writing: c_uchar,
    pub save_char: c_char,
    pub unused1: my_bool,
    pub unused2: my_bool,
    pub compress: my_bool,
    pub unused3: my_bool,
    pub unused: *mut c_uchar,
    pub last_errno: c_uint,
    pub error: c_uchar,
    pub unused4: my_bool,
    pub unused5: my_bool,
    pub last_error: [c_char; 512],
    pub sqlstate: [c_char; 6],
    pub extension: *mut c_void,

}

Fields
vio: *mut Viobuff: *mut c_ucharbuff_end: *mut c_ucharwrite_pos: *mut c_ucharread_pos: *mut c_ucharfd: my_socketremain_in_buf: c_ulonglength: c_ulongbuf_length: c_ulongwhere_b: c_ulongmax_packet: c_ulongmax_packet_size: c_ulongpkt_nr: c_uintcompress_pkt_nr: c_uintwrite_timeout: c_uintread_timeout: c_uintretry_count: c_uintfcntl: c_intreturn_status: *mut c_uintreading_or_writing: c_ucharsave_char: c_charunused1: my_boolunused2: my_boolcompress: my_boolunused3: my_boolunused: *mut c_ucharlast_errno: c_uinterror: c_ucharunused4: my_boolunused5: my_boollast_error: [c_char; 512]

Client library error message buffer. Actually belongs to struct MYSQL.

sqlstate: [c_char; 6]

Client library sqlstate buffer. Set along with the error message.

extension: *mut c_void

Extension pointer, for the caller private use. Any program linking with the networking library can use this pointer, which is handy when private connection specific data needs to be maintained. The mysqld server process uses this pointer internally, to maintain the server internal instrumentation for the connection.